Creación del nuevo trabajador Analista

+++ BuilderPatternTrabajador/Analista.java
@@ -4,7 +4,7 @@ public class Analista extends Trabajador {
     private double lineasCodigo;
     private final double VALOR_LINEA_JAVA = 1000;
     
-    public Analista(double salarioIntegral, double lineasCodigo) {
+    public Analista(double salarioIntegral, int lineasCodigo) {
         this.salarioIntegral = salarioIntegral;
         this.lineasCodigo = lineasCodigo;
     }
